+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `OxidEsales\EshopCommunity\Application\Controller\Admin\AdminController::isNewEditObject()`
   - `OxidEsales\EshopCommunity\Application\Model\Actions::isDefault()`
   - `OxidEsales\EshopCommunity\Core\Model\BaseModel::isPropertyLoaded()`
+  - `OxidEsales\EshopCommunity\Application\Controller\TextEditorHandler::disableTextEditor()`
+  - `OxidEsales\EshopCommunity\Application\Controller\TextEditorHandler::isTextEditorDisabled()`
+  - `OxidEsales\EshopCommunity\Application\Controller\Admin\AdminDetailsController::configureTextEditorHandler()`
+  - `OxidEsales\EshopCommunity\Application\Controller\Admin\AdminDetailsController::getTextEditorHandler()`
 
 ### Changed
 - In voucher series generation, if Coupon Number radio button checked, the number is marked as Required now. [PR-676](https://github.com/OXID-eSales/oxideshop_ce/pull/676)